refactor(transport): Move common test utility functions to fixtures
This de-duplicates code a little bit.
